implement parliament domain
